get config names from /config-names instead of /parquest-and-dataset-info for dataset-level runners
+ move runners to directories of their level
