+++ services/worker/src/worker/job_runner_factory.py
@@ -12,6 +12,8 @@ from libcommon.storage import StrPath
 from worker.config import AppConfig, FirstRowsConfig, ParquetAndDatasetInfoConfig
 from worker.job_runner import JobRunner
 from worker.job_runners.config.info import ConfigInfoJobRunner
+from worker.job_runners.config.parquet import ConfigParquetJobRunner
+from worker.job_runners.config.size import ConfigSizeJobRunner
 from worker.job_runners.config.split_names_from_dataset_info import (
     SplitNamesFromDatasetInfoJobRunner,
 )
@@ -19,17 +21,15 @@ from worker.job_runners.config.split_names_from_streaming import (
     SplitNamesFromStreamingJobRunner,
 )
 from worker.job_runners.config_names import ConfigNamesJobRunner
-from worker.job_runners.config_parquet import ConfigParquetJobRunner
-from worker.job_runners.config_size import ConfigSizeJobRunner
 from worker.job_runners.dataset.info import DatasetInfoJobRunner
+from worker.job_runners.dataset.parquet import DatasetParquetJobRunner
+from worker.job_runners.dataset.size import DatasetSizeJobRunner
 from worker.job_runners.dataset.split_names_from_dataset_info import (
     DatasetSplitNamesFromDatasetInfoJobRunner,
 )
 from worker.job_runners.dataset.split_names_from_streaming import (
     DatasetSplitNamesFromStreamingJobRunner,
 )
-from worker.job_runners.dataset_parquet import DatasetParquetJobRunner
-from worker.job_runners.dataset_size import DatasetSizeJobRunner
 from worker.job_runners.parquet_and_dataset_info import ParquetAndDatasetInfoJobRunner
 from worker.job_runners.split.first_rows_from_streaming import (
     SplitFirstRowsFromStreamingJobRunner,
