+++ services/worker/src/worker/job_runners/dataset/parquet.py
@@ -3,14 +3,15 @@
 
 import logging
 from http import HTTPStatus
-from typing import Any, Literal, Mapping, Optional, Tuple, TypedDict
+from typing import Any, List, Literal, Mapping, Optional, Tuple, TypedDict
 
-from libcommon.constants import PROCESSING_STEP_DATASET_SIZE_VERSION
+from libcommon.constants import PROCESSING_STEP_DATASET_PARQUET_VERSION
 from libcommon.dataset import DatasetNotFoundError
 from libcommon.simple_cache import DoesNotExist, SplitFullName, get_response
 
 from worker.job_runner import JobResult, JobRunner, JobRunnerError
-from worker.job_runners.config_size import ConfigSize, ConfigSizeResponse, SplitSize
+from worker.job_runners.config.parquet import ConfigParquetResponse
+from worker.job_runners.parquet_and_dataset_info import ParquetFileItem
 from worker.utils import PreviousJob
 
 SizesJobRunnerErrorCode = Literal[
@@ -19,27 +20,13 @@ SizesJobRunnerErrorCode = Literal[
 ]
 
 
-class DatasetSize(TypedDict):
-    dataset: str
-    num_bytes_original_files: int
-    num_bytes_parquet_files: int
-    num_bytes_memory: int
-    num_rows: int
-
-
-class DatasetSizeContent(TypedDict):
-    dataset: DatasetSize
-    configs: list[ConfigSize]
-    splits: list[SplitSize]
-
-
-class DatasetSizeResponse(TypedDict):
-    size: DatasetSizeContent
+class DatasetParquetResponse(TypedDict):
+    parquet_files: List[ParquetFileItem]
     pending: list[PreviousJob]
     failed: list[PreviousJob]
 
 
-class DatasetSizeJobRunnerError(JobRunnerError):
+class DatasetParquetJobRunnerError(JobRunnerError):
     """Base class for exceptions in this module."""
 
     def __init__(
@@ -55,29 +42,29 @@ class DatasetSizeJobRunnerError(JobRunnerError):
         )
 
 
-class PreviousStepStatusError(DatasetSizeJobRunnerError):
+class PreviousStepStatusError(DatasetParquetJobRunnerError):
     """Raised when the previous step gave an error. The job should not have been created."""
 
     def __init__(self, message: str, cause: Optional[BaseException] = None):
         super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepStatusError", cause, False)
 
 
-class PreviousStepFormatError(DatasetSizeJobRunnerError):
+class PreviousStepFormatError(DatasetParquetJobRunnerError):
     """Raised when the content of the previous step has not the expected format."""
 
     def __init__(self, message: str, cause: Optional[BaseException] = None):
         super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "PreviousStepFormatError", cause, False)
 
 
-def compute_sizes_response(dataset: str) -> Tuple[DatasetSizeResponse, float]:
+def compute_sizes_response(dataset: str) -> Tuple[DatasetParquetResponse, float]:
     """
-    Get the response of dataset-size for one specific dataset on huggingface.co.
+    Get the response of dataset-parquet for one specific dataset on huggingface.co.
     Args:
         dataset (`str`):
             A namespace (user or an organization) and a repo name separated
             by a `/`.
     Returns:
-        `DatasetSizeResponse`: An object with the sizes_response.
+        `DatasetParquetResponse`: An object with the parquet_response (list of parquet files).
     <Tip>
     Raises the following errors:
         - [`~job_runners.dataset_size.PreviousStepStatusError`]
@@ -86,35 +73,36 @@ def compute_sizes_response(dataset: str) -> Tuple[DatasetSizeResponse, float]:
             If the content of the previous step has not the expected format
     </Tip>
     """
-    logging.info(f"get sizes for dataset={dataset}")
+    logging.info(f"get parquet files for dataset={dataset}")
 
     try:
-        response = get_response(kind="/parquet-and-dataset-info", dataset=dataset)
+        response = get_response(kind="/config-names", dataset=dataset)
     except DoesNotExist as e:
-        raise DatasetNotFoundError(
-            "No response found in previous step for this dataset: '/parquet-and-dataset-info'.", e
-        ) from e
+        raise DatasetNotFoundError("No response for '/config-names' found for this dataset: .", e) from e
     if response["http_status"] != HTTPStatus.OK:
         raise PreviousStepStatusError(
-            f"Previous step gave an error: {response['http_status']}. This job should not have been created."
+            f"Previous step raised an error: {response['http_status']}. This job should not have been created."
         )
     content = response["content"]
+    if "config_names" not in content:
+        raise PreviousStepFormatError("'/config-names' did not return the expected content: 'config_names'.")
+
     try:
-        split_sizes: list[SplitSize] = []
-        config_sizes: list[ConfigSize] = []
+        parquet_files: list[ParquetFileItem] = []
         total = 0
         pending = []
         failed = []
-        for config in content["dataset_info"].keys():
+        for config_item in content["config_names"]:
+            config = config_item["config"]
             total += 1
             try:
-                response = get_response(kind="config-size", dataset=dataset, config=config)
+                response = get_response(kind="config-parquet", dataset=dataset, config=config)
             except DoesNotExist:
-                logging.debug("No response found in previous step for this dataset: 'config-size' endpoint.")
+                logging.debug("No response found in previous step for this dataset: 'config-parquet' endpoint.")
                 pending.append(
                     PreviousJob(
                         {
-                            "kind": "config-size",
+                            "kind": "config-parquet",
                             "dataset": dataset,
                             "config": config,
                             "split": None,
@@ -127,7 +115,7 @@ def compute_sizes_response(dataset: str) -> Tuple[DatasetSizeResponse, float]:
                 failed.append(
                     PreviousJob(
                         {
-                            "kind": "config-size",
+                            "kind": "config-parquet",
                             "dataset": dataset,
                             "config": config,
                             "split": None,
@@ -135,45 +123,31 @@ def compute_sizes_response(dataset: str) -> Tuple[DatasetSizeResponse, float]:
                     )
                 )
                 continue
-            config_size_content = ConfigSizeResponse(size=response["content"]["size"])
-            config_sizes.append(config_size_content["size"]["config"])
-            split_sizes.extend(config_size_content["size"]["splits"])
-        dataset_size: DatasetSize = {
-            "dataset": dataset,
-            "num_bytes_original_files": sum(config_size["num_bytes_original_files"] for config_size in config_sizes),
-            "num_bytes_parquet_files": sum(config_size["num_bytes_parquet_files"] for config_size in config_sizes),
-            "num_bytes_memory": sum(config_size["num_bytes_memory"] for config_size in config_sizes),
-            "num_rows": sum(config_size["num_rows"] for config_size in config_sizes),
-        }
+            config_parquet_content = ConfigParquetResponse(parquet_files=response["content"]["parquet_files"])
+            parquet_files.extend(config_parquet_content["parquet_files"])
     except Exception as e:
         raise PreviousStepFormatError("Previous step did not return the expected content.", e) from e
 
     progress = (total - len(pending)) / total if total else 1.0
 
     return (
-        DatasetSizeResponse(
-            {
-                "size": {
-                    "dataset": dataset_size,
-                    "configs": config_sizes,
-                    "splits": split_sizes,
-                },
-                "pending": pending,
-                "failed": failed,
-            }
+        DatasetParquetResponse(
+            parquet_files=parquet_files,
+            pending=pending,
+            failed=failed,
         ),
         progress,
     )
 
 
-class DatasetSizeJobRunner(JobRunner):
+class DatasetParquetJobRunner(JobRunner):
     @staticmethod
     def get_job_type() -> str:
-        return "dataset-size"
+        return "dataset-parquet"
 
     @staticmethod
     def get_job_runner_version() -> int:
-        return PROCESSING_STEP_DATASET_SIZE_VERSION
+        return PROCESSING_STEP_DATASET_PARQUET_VERSION
 
     def compute(self) -> JobResult:
         response_content, progress = compute_sizes_response(dataset=self.dataset)
@@ -182,6 +156,6 @@ class DatasetSizeJobRunner(JobRunner):
     def get_new_splits(self, content: Mapping[str, Any]) -> set[SplitFullName]:
         """Get the set of new splits, from the content created by the compute."""
         return {
-            SplitFullName(dataset=split_size["dataset"], config=split_size["config"], split=split_size["split"])
-            for split_size in content["size"]["splits"]
+            SplitFullName(dataset=parquet_file["dataset"], config=parquet_file["config"], split=parquet_file["split"])
+            for parquet_file in content["parquet_files"]
         }
